+++ pref/lib/src/page.dart
@@ -9,35 +9,55 @@ import 'service/base.dart';
 import 'service/cache.dart';
 import 'service/pref_service.dart';
 
-class PrefPage extends StatelessWidget {
+class PrefPage extends StatefulWidget {
   const PrefPage({
     @required this.children,
     this.cache = false,
-  }) : assert(children != null);
+  })  : assert(children != null),
+        assert(cache != null);
 
   final List<Widget> children;
 
   final bool cache;
 
-  Future<BasePrefService> _createCache(BasePrefService parent) async {
+  @override
+  _PrefPageState createState() => _PrefPageState();
+}
+
+class _PrefPageState extends State<PrefPage> {
+  Future<BasePrefService> _cache;
+
+  @override
+  void didChangeDependencies() {
+    if (widget.cache) {
+      _cache ??= _createCache();
+    }
+    super.didChangeDependencies();
+  }
+
+  Future<BasePrefService> _createCache() async {
     final service = PrefServiceCache();
-    await service.apply(parent);
+    await service.apply(_parent);
     return service;
   }
 
-  @override
-  Widget build(BuildContext context) {
-    // Check if we already have a BasePrefService
-    final _parent = PrefService.of(context);
+  BasePrefService get _parent {
+    final parent = PrefService.of(context);
 
-    if (_parent == null) {
+    // Check if we already have a BasePrefService
+    if (parent == null) {
       throw FlutterError(
           'No PrefService widget found in the tree. Unable to load settings');
     }
 
-    if (cache) {
+    return parent;
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    if (widget.cache) {
       return FutureBuilder<BasePrefService>(
-        future: _createCache(_parent),
+        future: _cache,
         builder: (BuildContext context, snapshot) {
           if (!snapshot.hasData) {
             return const SizedBox();
@@ -54,7 +74,7 @@ class PrefPage extends StatelessWidget {
                   await _parent.apply(_cache);
                   return true;
                 },
-                child: ListView(children: children),
+                child: ListView(children: widget.children),
               ),
             ),
           );
@@ -62,6 +82,6 @@ class PrefPage extends StatelessWidget {
       );
     }
 
-    return ListView(children: children);
+    return ListView(children: widget.children);
   }
 }