+++ pref/lib/src/title.dart
@@ -5,7 +5,10 @@
 
 import 'package:flutter/material.dart';
 
+/// A single fixed-height row that typically contains some text as well as
+/// an optional leading icon.
 class PrefTitle extends StatelessWidget {
+  /// Creates a preference title.
   const PrefTitle({
     @required this.title,
     this.decoration,
@@ -16,16 +19,26 @@ class PrefTitle extends StatelessWidget {
     this.padding,
   }) : assert(title != null);
 
+  /// The primary content of the Title.
   final Widget title;
 
+  /// The decoration applied.
   final Decoration decoration;
 
+  /// A widget to display before the title.
   final Widget leading;
+
+  /// Additional content displayed below the title.
   final Widget subtitle;
 
+  /// Called when the user taps this title.
   final Function onTap;
 
+  /// Empty space to inscribe inside the [decoration]. The title, if any, is
+  /// placed inside this padding.
   final EdgeInsetsGeometry padding;
+
+  /// Empty space to surround the [decoration] and the Title.
   final EdgeInsetsGeometry margin;
 
   @override