+++ pref/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Cache() copies the default values
 - Add listen parameter to PerfService.of()
 - Fix cached Page and Dialog
+- Add more documentation
 
 ## 1.3.1